Intro level LSTM-RNN
So you want to play with neural networks, don't you?

Well first you have to get past all of the jargon; like what is a
feature, labels, what's a time step, what are patterns, what should a
sequence length be, how to pick a batch size, what even is input shape?

I'm just as confused as you are, but let's keep figuring this out
anyways